BUG: using InputPixelType to hold label count can overflow.
The member variable m_TotalLabelCount is set to the maximum
pixel value in the input images, plus one. For any unsigned pixel
type if the maximum input value equals
itk::NumericTraits<InputPixelType>::max(), then the TotalLabelCount
gets set to zero, and zero-length tables end up causing memory errors
and crashes.

+++ Modules/Segmentation/LabelVoting/include/itkLabelVotingImageFilter.hxx
@@ -71,7 +71,8 @@ LabelVotingImageFilter< TInputImage, TOutputImage >
   Superclass::BeforeThreadedGenerateData();
 
   // determine the maximum label in all input images
-  this->m_TotalLabelCount = this->ComputeMaximumInputValue() + 1;
+  this->m_TotalLabelCount =
+    static_cast<size_t>(this->ComputeMaximumInputValue()) + 1;
 
   if ( !this->m_HasLabelForUndecidedPixels )
     {
